requirements.txt
beautifulsoup4==4.9.1
google==3.0.0
google_search==1.0.2
googlesearch-python==2020.0.2
email-validator==1.1.2
Flask_Bootstrap==3.3.7.1
Flask_Moment==0.10.0
Werkzeug==1.0.1
click==7.1.2
alembic==1.4.2
Flask_Babel==1.0.0
Flask_Login==0.5.0
Flask_WTF==0.14.3
SQLAlchemy==1.3.20
WTForms==2.3.1
Flask_Mail==0.9.1
Flask==1.1.2
Flask_SQLAlchemy==2.4.4
requests==2.24.0
Flask_Migrate==2.5.3
PyJWT==2.0.1
